file restructure & fixes to ping.py
